refactor(drawer): move logic to brain, keep helm styles-only
Addresses @MerlinMoos' review feedback (#1326): "keep hlm files small as
possible only with styles — split all related logic to the brn package
... people can use the brn without using our hlm."

Brain now owns the full behaviour:

- `BrnDrawer` — coordinator directive holding the state machine
  (`closed` / `opening` / `open` / `closing`), signals, inputs, outputs,
  and all effects (gesture attachment, resize observation, scroll lock,
  keyboard avoidance, modal effect).
- `BrnDrawerContent` / `BrnDrawerHandle` / `BrnDrawerScroller` /
  `BrnDrawerFooter` / `BrnDrawerOverlay` — thin sub-directives that
  register their element refs with the coordinator and surface
  dynamic host bindings (transform, opacity, touch-action, etc.).

Helm is now pure presentation:

- `HlmDrawer` is a component with `hostDirectives: [BrnDrawer]` that
  forwards all inputs/outputs and applies Spartan semantic Tailwind
  tokens (`bg-background`, `text-foreground`, `bg-muted-foreground`,
  `border-border`) to the template elements via class attributes.
- `HlmDrawerHeader` / `HlmDrawerFooter` / `HlmDrawerTitle` /
  `HlmDrawerDescription` switched from the verbose
  `injectCustomClassSettable + computed + effect` boilerplate to the
  single-line `classes(() => '...')` helper, matching `HlmSheet*` style.

Consumers can now pick `@spartan-ng/brain/drawer` alone and bring their
own styling. `@use-gesture/vanilla` moved from helm's peerDeps to
brain's (as an optional peer, same pattern as `luxon` for
`brain/date-time`).

Behaviour unchanged: all 18 brain unit tests still pass, the
generate-ui-docs snapshot picks up the new `BrnDrawer*` primitives,
Storybook builds cleanly.

Refs #118
