+++ libs/brain/drawer/src/index.ts
@@ -1,3 +1,17 @@
+import { BrnDrawer } from './lib/brn-drawer';
+import { BrnDrawerContent } from './lib/brn-drawer-content';
+import { BrnDrawerFooter } from './lib/brn-drawer-footer';
+import { BrnDrawerHandle } from './lib/brn-drawer-handle';
+import { BrnDrawerOverlay } from './lib/brn-drawer-overlay';
+import { BrnDrawerScroller } from './lib/brn-drawer-scroller';
+
+export { BrnDrawer } from './lib/brn-drawer';
+export { BrnDrawerContent } from './lib/brn-drawer-content';
+export { BrnDrawerFooter } from './lib/brn-drawer-footer';
+export { BrnDrawerHandle } from './lib/brn-drawer-handle';
+export { BrnDrawerOverlay } from './lib/brn-drawer-overlay';
+export { BrnDrawerScroller } from './lib/brn-drawer-scroller';
+
 export {
 	BRN_DRAWER_DEFAULT_DRAG_CLOSE_THRESHOLD,
 	BRN_DRAWER_DEFAULT_DRAG_VELOCITY_THRESHOLD,
@@ -17,3 +31,12 @@ export type { BrnDrawerDetent, BrnDrawerSnapPoint, BrnDrawerState, BrnDrawerTwee
 export { isAscendingOrder, isIOS } from './lib/utils/platform';
 export { computeSnapPoints, handleHighVelocityDrag, handleLowVelocityDrag } from './lib/utils/snap';
 export { tweenTo } from './lib/utils/tween';
+
+export const BrnDrawerImports = [
+	BrnDrawer,
+	BrnDrawerContent,
+	BrnDrawerOverlay,
+	BrnDrawerHandle,
+	BrnDrawerScroller,
+	BrnDrawerFooter,
+] as const;
