+++ libs/helm/drawer/src/lib/hlm-drawer-header.directive.ts
@@ -1,25 +1,14 @@
-import { computed, Directive, effect, input, untracked } from '@angular/core';
-import { injectCustomClassSettable } from '@spartan-ng/brain/core';
-import { hlm } from '@spartan-ng/helm/utils';
-import type { ClassValue } from 'clsx';
+import { Directive } from '@angular/core';
+import { classes } from '@spartan-ng/helm/utils';
 
 @Directive({
 	selector: '[hlmDrawerHeader]',
 	host: {
-		'[class]': '_computedClass()',
+		'data-slot': 'drawer-header',
 	},
 })
 export class HlmDrawerHeader {
-	private readonly _classSettable = injectCustomClassSettable({ optional: true, host: true });
-	public readonly userClass = input<ClassValue>('', { alias: 'class' });
-	protected readonly _computedClass = computed(() =>
-		hlm('flex flex-col gap-1.5 p-4 text-center sm:text-left', this.userClass()),
-	);
-
 	constructor() {
-		effect(() => {
-			const classValue = this._computedClass();
-			untracked(() => this._classSettable?.setClassToCustomElement(classValue));
-		});
+		classes(() => 'flex flex-col gap-1.5 p-4 text-center sm:text-left');
 	}
 }
