+++ libs/helm/drawer/src/lib/hlm-drawer-title.directive.ts
@@ -1,25 +1,14 @@
-import { computed, Directive, effect, input, untracked } from '@angular/core';
-import { injectCustomClassSettable } from '@spartan-ng/brain/core';
-import { hlm } from '@spartan-ng/helm/utils';
-import type { ClassValue } from 'clsx';
+import { Directive } from '@angular/core';
+import { classes } from '@spartan-ng/helm/utils';
 
 @Directive({
 	selector: '[hlmDrawerTitle]',
 	host: {
-		'[class]': '_computedClass()',
+		'data-slot': 'drawer-title',
 	},
 })
 export class HlmDrawerTitle {
-	private readonly _classSettable = injectCustomClassSettable({ optional: true, host: true });
-	public readonly userClass = input<ClassValue>('', { alias: 'class' });
-	protected readonly _computedClass = computed(() =>
-		hlm('text-foreground text-lg leading-none font-semibold tracking-tight', this.userClass()),
-	);
-
 	constructor() {
-		effect(() => {
-			const classValue = this._computedClass();
-			untracked(() => this._classSettable?.setClassToCustomElement(classValue));
-		});
+		classes(() => 'text-foreground text-lg leading-none font-semibold tracking-tight');
 	}
 }
