+++ libs/helm/package.json
@@ -11,7 +11,6 @@
 		"@ng-icons/core": ">=32.0.0 <34.0.0",
 		"@ng-icons/lucide": ">=32.0.0 <34.0.0",
 		"@spartan-ng/brain": "0.0.1-alpha.675",
-		"@use-gesture/vanilla": "^10.3.1",
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^2.1.1",
 		"embla-carousel-angular": "^20.0.0",
